chore: cleanup server utils folders + files
