+++ app/lib/server-utils/seo.server.ts
@@ -6,6 +6,7 @@ import type {
   Image,
   Shop,
 } from '@shopify/hydrogen/storefront-api-types';
+import type {AppLoadContext} from '@shopify/remix-oxygen';
 import type {CollectionPage, Offer, SearchResultsPage} from 'schema-dts';
 
 import type {
@@ -16,7 +17,31 @@ import type {
   Seo,
 } from '~/lib/types';
 
+import {getShop, getSiteSettings} from './settings.server';
+
 type SeoMedia = SeoConfig['media'];
+
+export const getAccountSeo = async (
+  context: AppLoadContext,
+  accountTitle: string,
+) => {
+  const [shop, siteSettings] = await Promise.all([
+    getShop(context),
+    getSiteSettings(context),
+  ]);
+  const {title: seoSiteTitle} = {
+    ...siteSettings?.data?.siteSettings?.seo,
+  } as Seo;
+  const noIndex =
+    !!siteSettings?.data?.siteSettings?.settings?.account?.noIndex;
+  const noFollow =
+    !!siteSettings?.data?.siteSettings?.settings?.account?.noFollow;
+  const robots = {noIndex, noFollow};
+  const siteTitle = seoSiteTitle || shop?.name || '';
+  const title = `${accountTitle} | ${siteTitle}`;
+  const seo = {title, robots};
+  return seo;
+};
 
 const getMeta = ({
   affixSiteTitleToSeoTitle = true,
