+++ app/lib/utils/index.ts
@@ -1,6 +1,6 @@
 export * from './document.utils';
+export * from './general.utils';
 export * from './locale.utils';
-export * from './oxygen.utils';
+export * from './metafields.utils';
 export * from './pack.utils';
 export * from './product.utils';
-export * from './server.utils';
