+++ app/routes/($locale).account.addresses.tsx
@@ -7,7 +7,7 @@ import type {
 import {AnalyticsPageType, getSeoMeta} from '@shopify/hydrogen';
 
 import {customerAddressesAction} from '~/lib/customer/addresses.server';
-import {getAccountSeo} from '~/lib/utils';
+import {getAccountSeo} from '~/lib/server-utils/seo.server';
 import {CustomerAccountLayout} from '~/components/AccountLayout/CustomerAccountLayout';
 import {Addresses} from '~/components/Account/Addresses/Addresses';
