+++ app/routes/($locale).account.profile.tsx
@@ -6,7 +6,7 @@ import type {
 } from '@shopify/remix-oxygen';
 import {AnalyticsPageType, getSeoMeta} from '@shopify/hydrogen';
 
-import {getAccountSeo} from '~/lib/utils';
+import {getAccountSeo} from '~/lib/server-utils/seo.server';
 import {CustomerAccountLayout} from '~/components/AccountLayout/CustomerAccountLayout';
 import {Profile} from '~/components/Account/Profile';
 import {customerUpdateProfileAction} from '~/lib/customer/profile.server';
