+++ app/routes/($locale).products.$handle[.]json.tsx
@@ -1,9 +1,9 @@
 import type {LoaderFunctionArgs} from '@shopify/remix-oxygen';
 
-import {getSelectedProductOptions} from '~/lib/products.server';
+import {getSelectedProductOptions} from '~/lib/server-utils/product.server';
 import {PRODUCT_QUERY} from '~/data/graphql/storefront/product';
 import {routeHeaders} from '~/data/cache';
-import {transformShopifyGids} from '~/lib/utils/server.utils';
+import {transformShopifyGids} from '~/lib/utils';
 export const headers = routeHeaders;
 
 export async function loader({params, context, request}: LoaderFunctionArgs) {
